refactor all event scraping to background tasks
this was a big one. this change moves all event scraping from being
driven by user HTTP requests to tasks that run in the background of the
process at fixed intervals that persist events to the DB, with the HTTP
requests becoming simply a means to query the DB. It's now much simpler
to create a new type of scraper, you must only implement the UrlScraper
interface and register it. There is a framework that handles logging,
retries, and error handling for you.

TODO:
- with-friends and wix still need some work, i don't currently support
POST requests or non-unique URLs
- image scraping is still to-do
- store a log of scrapes in the db and surface that somehow?
- enable the eventbrite single scraper, maybe allowing promoters to send
us events directly without going through github?
- add an arch diagram to the readme/docs somewhere for new devs