serialize instagram scraping, give up on rate limit error
It doesn't actually take that long to run the Instagram scrape workflow, so
moving to a serial execution model isn't going to cause any massive issues for
us. By moving to a serial execution model, we gain the ability to give up on the
scrape flow if we get a rate limit exception from Instagram. This ensures that
as soon as we get rate limited, we stop making the rate limit exception worse.