delete instagram-specific data model
the previous data model made an explicit distinction between instagram
events and events not from instagram. there was really no need to do
this, the urlevent schema was perfectly capable of storing instagram
event details. this schism was initially made because there's a very
weird way we have to scrape instagram events, and i refactorred that
_before_ creating the urlevent data model. after creating that, i
realized i really should unify everything to be based on it. this commit
does that unification, and as a benefit it fixes image scraping for
non-instagram images.

there is now _one_ event type that is used everywhere. the only thing
that knows instagram is a fucked-up mess is the scraping system, as it
should be. the frontend is now entirely event-agnostic, so features no
longer have to be implemented twice. thank god.