Move to server-side filtering
The current client-side filtering requires us to send all event data
down to the client. This is a lot of data to send over the network and
makes the initial page load slower. This change moves the filtering to
the server side and (most importantly) only returns events for the
current month.